+++ src/main/java/com/github/theredbrain/spellengineextension/spell_engine/DuckSpellModifierMixin.java
@@ -1,0 +1,25 @@
+package com.github.theredbrain.spellengineextension.spell_engine;
+
+public interface DuckSpellModifierMixin {
+
+	float spellengineextension$getAdditionalHealthCost();
+
+	void spellengineextension$setAdditionalHealthCost(float additionalHealthCost);
+
+	float spellengineextension$getAdditionalManaCost();
+
+	void spellengineextension$setAdditionalManaCost(float additionalManaCost);
+
+	float spellengineextension$getAdditionalStaminaCost();
+
+	void spellengineextension$setAdditionalStaminaCost(float additionalStaminaCost);
+
+	double spellengineextension$getAdditionalDirectDamage();
+
+	void spellengineextension$setAdditionalDirectDamage(double additionalDirectDamage);
+
+	double spellengineextension$getAdditionalDirectHealing();
+
+	void spellengineextension$setAdditionalDirectHealing(double additionalDirectHealing);
+
+}